🎨 Palette: Add toast notifications for long operations
💡 What: Added `gr.Info()` toasts for PDF conversion and model download operations. Also fixed failing test assertions regarding the generator raised exceptions.
🎯 Why: Text-only terminal logs in Gradio applications are often insufficient for conveying critical and terminal state information because they require users to constantly monitor the scrolling area. Added toasts to provide immediate visible feedback.
📸 Before/After: N/A (toast notification)
♿ Accessibility: Ensures that screen readers or users not focused on the log area are alerted when long processes start.

+++ scripts/convert_one.py
@@ -72,7 +72,7 @@ def _event_printer() -> callable:
         prefix = f"[{event.phase.upper():<10s}] t={elapsed:6.1f}s"
         bar = ""
         if event.total and event.percent:
-            filled = int(round(event.percent / 5))
+            filled = round(event.percent / 5)
             bar = f" [{('█' * filled):<20s}] {event.percent:5.1f}%"
         print(f"{prefix}{bar} {event.message}", flush=True)
         if event.phase == "error":
